requirements.txt
mysql-connector-python==9.4.0
PyYAML==6.0.2
